+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+* upgrade to [actix-web](https://github.com/actix/actix-web) version 0.7.2
+* restructure lib
+* [clippy](https://github.com/rust-lang-nursery/rust-clippy) compliance
+* add one more test
+
 ## 0.2.3
 * deliver request body
 
@@ -12,10 +18,10 @@
 * fix visibility of received request
 
 ## 0.2.0
-* reimplement server with actix/actix-web
+* reimplement server with [actix-web](https://github.com/actix/actix-web)
 
 ## 0.1.1
 * fix retreiving of requests 
 
 ## 0.1.0
-* first version based on iron/iron
+* first version based on [iron](https://github.com/iron/iron)